Implement a range table for the quantization matricies in the setup header. We still use (and only support in decode) the original set of three VP3 matricies; decoder support will be forthcoming.
THIS IS AN INCOMPATIBLE BITSTREAM CHANGE.

Lossless reformatting is possible, as with the other header changes.

Under the new scheme, different matricies can be specified for each of six different kinds of
fragments: Intra or Inter Y, U, and V, and for each quality level. To save space, explicit
matricies can be given only for specific q values and are linearly interpolated between.
Likewise, a fragment category can be marked the same as the previous category and inter fragments
can be marked the same as the corresponding intra plane. This adds only a few bytes to the
previous header format in the VP3 case, but adds significant flexibility for future encoders.