Much debug code added. Discovered source of latency in seeking, about to start fixing. Latency is due to a thread syncronisation error between the application thread and the streaming thread, which under appropriate conditions after a seek are out of synch.
The result is that the demux starts sending new smaples before the streaming thread has comlpeted flushing and returned to a ready state. This causes all delivery mechanisms to reject delivery of samples (like what happens when you stop, all following samples are rejected). The result is that the demux is working off the early timebase but the decoders don't see these early samples, when they finally receive samples, which have a later timestamp, they resync to them.

Need to ensure that all flush operations are completed before resuming the flow of samples.