Issue an explicit warning if the docs can't be built at 'make dist' time, as is now required.
This is a little less broken than the previous config in that the dist 
will need to have the build docs, and a normal build from a dist tarball
won't try. However if you --enable-docs on the dist tarball, it will 
rebuild them even if they're up to date. Among other things, this makes
'make distcheck' annoying.