Added part of the FLAC homepage.  I'm not sure about the directory name, but I know know what else to call it.
The auto-negotion works now so that if a person's browser is set to use Russian, it will auto-load the .ru page if it's available.  We still need a way to indicate that there are alternate translations for pages for those that aren't using good browsers, and for those that would like to switch between translations.