Commit the built specification to the repository.
We have traditionally avoided putting generated files in the 
respository; it's a waste of space and makes for useless diffs.
However, the docbook tool situation still hasn't improved to
the point where the docs build out of the box for most people.

The real impetus was a short-term hack to simplify the website
reorg; if we ever do get things building cleanly, the server
should do that before pushing to the mirrors.