MAJOR UPDATE
Commit Ivo Concalves rewrite of the HTML.  This is a *large* update to do such things as convert tables to structural HTML.  

Finally getting a copyright statement in there.  Used Creative Commons Attribution-NoDerivs 2.5 License because it is so much less
cumbersome than the GFDL, and because the GFDL requires us to allow derivatives.

Improve explanation of link, meta and extension elements.

Check links and remove a couple broken internal links in the TOC.

Removed Dave Brown from authors list and added him to acknowledgements.  It just seems patently false to claim that he contributed
to v1.  

Updated acknowledgements section.

Deleted usecases section, which is mainly a source of bloat and potential errors and doesn't belong in a standards document.

Changed "2.3.2 Requirements notation" to say that [RFC2119] is the definition of MUST, SHOULD, etc.  The old language was 
confusing.  

***I do not believe that any of these edits will invalidate any real world implementation of the specification.***