Update XCode project for recent source file changes.
Also removes the prebind flags, which throw a warning when linked 
against the 10.4 or later SDKs, as we do by default. There are
still some warnings, so there must be an additional setting for
it I didn't find.

Replace an instance of the old USE_ASM cpp define in the Theora target 
with the new OC_X86_ASM.