Revert monospace font change on download listings.
SVN r18267 also set a font-family:monospace on
the download list <td> elements. This is better
done with css, but the font choice also fails
to harmonize with the rest of the site design.

It makes some sense for the checksums to be
in a monospaced font for easier visual comparision,
but not the whole line.